requirements.txt
# Requirements for Weekly Market Report System

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Data Fetching
yfinance>=0.2.36
futu-api>=9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Browser Automation (Playwright)
playwright>=1.40.0

# Notion Integration
notion-client>=2.2.0

# Social Media APIs
tweepy>=4.14.0          # X (Twitter) API

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
tabulate>=0.9.0         # 表格格式化

# Chart Screenshots
Pillow>=10.0.0

# Scheduling
schedule>=1.2.0
